requirements.txt
numpy
opencv-python
pillow
pandas
scipy
matplotlib
prompt_toolkit
torch
torchvision
tqdm